+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.3
+### Fix
+* Using AWS SDK defaultProvider to get credentials in a more consisten manner
+  
+## 0.1.2
+### Fix
+* Returning undefined credentials in case passed profile does not exist
+
 ## 0.1.1
 ### Chore
 * Added credits to Readme